backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
transformers
torch
sentencepiece
pdfplumber
python-docx
